+++ Microsoft.DotNet.Interactive/Display.cs
@@ -1,0 +1,44 @@
+﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+using System.Threading.Tasks;
+using Microsoft.DotNet.Interactive.Commands;
+using Microsoft.DotNet.Interactive.Rendering;
+
+namespace Microsoft.DotNet.Interactive
+{
+    public class Display : IDisplay
+    {
+        private readonly string _displayId;
+        private readonly string _mimeType;
+
+        public Display(string displayId, string mimeType)
+        {
+            if (string.IsNullOrWhiteSpace(displayId))
+            {
+                throw new ArgumentException("Value cannot be null or whitespace.", nameof(displayId));
+            }
+
+            if (string.IsNullOrWhiteSpace(mimeType))
+            {
+                throw new ArgumentException("Value cannot be null or whitespace.", nameof(mimeType));
+            }
+            _displayId = displayId;
+            _mimeType = mimeType;
+        }
+
+        public void Update(object updatedValue)
+        {
+            var formatted = new FormattedValue(
+                _mimeType,
+                updatedValue.ToDisplayString(_mimeType));
+
+            var kernel = KernelInvocationContext.Current.Kernel;
+
+            Task.Run(() =>
+                    kernel.SendAsync(new UpdateDisplayedValue(formatted, _displayId)))
+                .Wait();
+        }
+    }
+}
