+++ Microsoft.DotNet.Interactive/Kernel.cs
@@ -28,39 +28,7 @@ namespace Microsoft.DotNet.Interactive
             return new Display(displayId,  mimeType);
         }
 
-        private class Display : IDisplay
-        {
-            private readonly string _displayId;
-            private readonly string _mimeType;
-
-            public Display(string displayId, string mimeType)
-            {
-                if (string.IsNullOrWhiteSpace(displayId))
-                {
-                    throw new ArgumentException("Value cannot be null or whitespace.", nameof(displayId));
-                }
-
-                if (string.IsNullOrWhiteSpace(mimeType))
-                {
-                    throw new ArgumentException("Value cannot be null or whitespace.", nameof(mimeType));
-                }
-                _displayId = displayId;
-                _mimeType = mimeType;
-            }
-
-            public void Update(object updatedValue)
-            {
-                var formatted = new FormattedValue(
-                    _mimeType,
-                    updatedValue.ToDisplayString(_mimeType));
-
-                var kernel = KernelInvocationContext.Current.Kernel;
-
-                Task.Run(() =>
-                        kernel.SendAsync(new UpdateDisplayedValue(formatted, _displayId)))
-                    .Wait();
-            }
-        }
+      
 
         public static void Javascript(
             string scriptContent)
